requirements.txt
Flask==1.1.2
SQLAlchemy~=1.4.6
Flask_Migrate==2.7.0
Flask_SQLAlchemy==2.5.1
python-dotenv==0.19.2

requests~=2.26.0